Install less documentation, clean more files.
INSTALL is of not useful outside the source tree.
intltool-*.in were probably installed by error.
(is src/lingot-logo.xpm useful?)

intltool-*.in and po/Makefile.in.in may be removed by distclean.
DISTCLEANFILES only considers files, not directories.

In README, tell the user to create m4/, else intltoolize fails.