Improve setting and transmission of build flags.
Do not repeat the detail of ./configure --help in README.md.

AC_ARG_WITH is recommended instead of AC_ARG_ENABLE for external
dependencies.  The choices in ./configure are now
* if available (the default)
* explicit yes (fail if unavailable)
* explicit no

Directly changing user variables like CFLAGS from configure.ac or
Makefile.am causes a lot of problems. Also, in automake, CFLAGS is
always appended, and AM_CFLAGS can be ignored nowadays in favor of
lingot_CFLAGS.

Only run the tests if libcunit is available.

Pass each l_CFLAGS l_LIBS for consistency, even if some happen to be
empty for now (like OSS_CFLAGS or PULSEAUDIO_CFLAGS).

On the other hand, the exemples do really not need GTK.

The -pthread option is intended for both cflags and ldflags.